refactor: move LuaJIT enums to dedicated `Enums` namespace and update imports accordingly
