+++ src/Extensions/NeoServer.Scripts.LuaJIT/Enums/MagicEffect.cs
@@ -1,4 +1,4 @@
-namespace NeoServer.Scripts.LuaJIT.Models;
+namespace NeoServer.Scripts.LuaJIT.Enums;
 
 public enum MagicEffect : ushort
 {
