+++ src/Extensions/NeoServer.Scripts.LuaJIT/Models/Callbacks/ValueCallback.cs
@@ -1,6 +1,7 @@
 using LuaNET;
 using NeoServer.Domain.Common;
 using NeoServer.Domain.Common.Contracts.Creatures;
+using NeoServer.Scripts.LuaJIT.Enums;
 
 namespace NeoServer.Scripts.LuaJIT.Models.Callbacks;
